Don't demodulate the current slot's entire guard interval.
We now attempt to demodulate an entire slot's worth of data starting and half a
guard interval before the start of the slot. Before we were demodulating the
entire current slot plus half the previous guard interval, which led to
duplicate packets if nodes' clocks were misaligned in just the right way.