Separate TX PHY parameters into separate TXParam type.
TX PHY parameters (checksum, FEC, and modulation scheme) are now kept in
a separate, distinct object. Individual nodes reference a TXParam.

Additionally, 0dBFS can now be calculated on a per-TXParam basis. We use a new
generic estimator framework to update an average. Two gains are therefore
applied to each packet: a per-TXParam gain, which aims to hit 0 dBFS, and a
per-node TX power control gain, which defaults to 1.