Always restart the retransmission timer when we receive a new ACK.
Before we were only canceling the retransmission timer if ALL packets had been
ACK'ed. This meant that if there was a constant stream of data, the
retransmission timer would always fire.

Now we always cancel the retransmission timer when we receive an ACK, and if we
still have more un-ACK'ed data, we also restart the retransmission timer.