Set python debug level in standalone radio scripts.
Before this patch, the `-v` switch would only turn on verbose messages in the
C++ part of the radio. Now it does that and additionally turns on verbose mode
in the python logger. The `-d` switch turns on verbose messages in the
C++ part of the radio and debug logging in the python code.