Add a separate parameter to control amount of overlap during demod.
During demodulation, we demodulate a portion of the previous slot before
demodulating the current slot to account for clock misalignment between
nodes. We used to demodulate one-half of the guard interval at the tail end of
the previous slot, which was fine when we were using TDMA. But when we use FDMA,
we may want a very small guard interval *even if* clocks are not aligned at that
granularity. Therefore, we now separately specify the amount of overlap (in
seconds) during demodulation.