Packets without data get the broadcast TX params.
Why? We have no way to tell whether or not a data-less packets was sucessfully
recevied, since it won't be ACK'ed! So, if we happen to use a problematic MCS,
we'll never know.

Using a lower-bit-rate MCS isn't a big deal, because if we were actually
transmitting data, our ACK's would be piggy-backed on a data packet.