The slot queue is now local to slotted MAC's txWorker thread.
This allows us to avoid using any kind of lock on the queue since it is local to
txWorker. It also simplifies the code IMHO. Finally, it seems to miss TX slots
slightly less frequently which was the hope (due to the removal of the lock on
the slot queue).