Fix channel iteration in channelizers.
The logic for parallel channelization was incorrect...doh! We need to initialize
the current channel with the thread's ID (tid), and then increment it by *the
number of threads* on each iteration, not the number of channels!

Also, a given thread now only reconfigure once before attempting to demodulate
all channels for which it is responsible.