Correctly set snapshot offset for final snapshot slot.
When collecting snapshots, we must collect self transmission information beyond
the end of the snapshot in order to catch transmissions that begin during the
snapshot but do not end until the snapshot is over---we don't detect
transmissions until they are complete.

For the time-domain channelizer, we must collect self-transmissions for one
extra slot. For the frequency-domain channelizer, we must collect
self-transmissions for two extra slots since time-domain data may be pushed to
the following frequency-domain FFT buffer.

The gist of the problem is that we weren't correctly computing the snapshot
offset of this final extra slot, so we were missing these final
transmissions---we were computing the offset of the next slot using the size of
the current IQ buffer, which may not have been valid because it was in the
process of being received. We also now collect two extra slots' worth of
self-transmissions for the frequency-domain channelizer.