Properly compute gain to compensate for multichannel transmission.
We pessimistically assume that if there are N channels we can transmit on
during *some* slot, then we need to apply a 1/N compensation factor to modulated
signal. Since gain must be the same across slots, and because superslots may be
enabled, we are pessimistic in order to simplify the gain calculation.