Programmatically add ARP table entries for traffic interface.
We now directly add static ARP table entries for the traffic interface instead
of shelling out to arp, and we add the entries immediately after creating the
radio instead of waiting until we are given the start command. The asyncio
coroutine that was responsible for this was overloading the neighbor discovery
coroutine; ARP table stuffing should be much more robust now.