Don't immediately append feedback to ACK and NAK packets.
Instead of immediately appending feedback to ACK and NAK packets, we set a flag
indicating the packet needs selective ACK feedback attached. If we attach
feedback immediately, by the time the packet is sent, the feedback can be out of
date.

ACK and NAK packets should always have selective ACK feedback attached to
increase the accuracy of PER estimates.