Support both slot_timestamp and timestamp for received packets.
For received packets, the slot_timestamp column contains the timestamp of the
slot and the timestamp column contains the timestamp at which the packet was
received.

The name of the timestamp column was changed to slot_timestamp in the logger, so
if the slot_timestamp column doesn't exist, we rename timestamp to
slot_timestamp and proceed. The timestamp column is calculated from the
slot_timestamp, start sample offset, and bandwidth.