Don't mark MCS entries invalid in the MAC.
MCS entries no longer have a valid flag. Our original strategy was to see if the
largest possible packet could be sent by the current MAC at a given MCS, and if
not, mark it as invalid. Now the PHY just drops the packet and emits a warning
if and when a modulated packet is larger than a full slot.